refactor: InputViewWeb, OutputViewWeb을 컴포넌트 단위로 리팩토링
